+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 ### Fixed
 - Using key for set value in enum fields.
+- Missing attributes in `audit-entity-offline` serializer.
 
 ## [0.5.0-beta.3] - 2016-08-24
 ### Added