+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change Log
 
-## [Unreleased]
+## [0.5.0-beta.3] - 2016-08-25
 ### Fixed
 - Using captions instead keys for enum fields in `syncer`.
 - Missing attributes in `audit-entity-offline` serializer.